Extract large test modules in storage crate to separate files
