+++ .jules/atlas.md
@@ -9,3 +9,6 @@
 ## 2026-04-11 - Module Coupling Refactor
 **Tangle:** Several circular dependencies existed across modules: `types` <-> `values` in `relvar-core`, `mvcc` <-> `storage`, and `wal` <-> `storage` in `relvar-storage`.
 **Blueprint:** Refactored dependencies by reallocating shared functionality. The `selector` method on `ScalarType` was replaced with `ScalarValue::select(&type, value)` mapping to correct responsibility since `ScalarValue` relies on `types`, thereby untangling the `relvar-core` loop. In `relvar-storage`, `collect_garbage` was relocated directly into the `storage::heap` namespace, eliminating the cyclical jump between `mvcc` and `storage`. Also decoupled `wal` <-> `storage` by replacing domain-specific struct pointers in `WalRecord` with primitive `u64` representing the page id, establishing unidirectional flow and solid boundaries.
+## 2025-04-16 - [De-bloat storage modules]
+**Tangle:** The `relvar-storage/src/storage/heap.rs` and `relvar-storage/src/persistent_engine.rs` modules were growing extremely large (4.5k+ and 1.8k+ lines respectively), causing a "Blob" anti-pattern due to having massive inline test modules.
+**Blueprint:** Extracted the massive test modules into their own files (`relvar-storage/src/storage/heap/heap_tests.rs` and `relvar-storage/src/persistent_engine/persistent_engine_tests.rs`) to separate test code from core logic, greatly reducing the bloat and improving readability.
